+++ packages/client/src/views/Settings/SettingsPage.tsx
@@ -13,7 +13,7 @@ import * as React from 'react'
 import { useIntl } from 'react-intl'
 import styled from '@client/styledComponents'
 import { Header } from '@client/components/interface/Header/Header'
-import { ListViewSimplified } from '@opencrvs/components/lib/interface/ListViewSimplified/ListViewSimplified'
+import { ListViewSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import { userMessages as messages } from '@client/i18n/messages'
 import { Navigation } from '@client/components/interface/Navigation'
 import { Content } from '@opencrvs/components/lib/Content'
