+++ packages/client/src/views/SysAdmin/Config/Application/Application.tsx
@@ -16,7 +16,7 @@ import { Content } from '@opencrvs/components/lib/Content'
 import { messages } from '@client/i18n/messages/views/config'
 import styled from 'styled-components'
 import { FormTabs } from '@opencrvs/components/lib/FormTabs'
-import { ListViewSimplified } from '@opencrvs/components/lib/interface/ListViewSimplified/ListViewSimplified'
+import { ListViewSimplified } from '@opencrvs/components/lib/ListViewSimplified'
 import {
   ApplicationName,
   Currency,
