+++ packages/components/src/interface/index.ts
@@ -15,7 +15,6 @@ export * from './SectionDrawer'
 export * from './SearchTool'
 export * from './Header/AppHeader'
 export * from './Header/PageHeader'
-export * from './ViewData'
 export * from './EventTopBar'
 export * from './ListTable'
 export * from './TableView'
